Update deviceList.js
The table in the database is changed from devices to deviceCheckResult.
Each record is one assessment check result.  It has a state, and a message.  This is displayed on the device list.  That is all this console needs to do.
Bypass shuffle.
Bypass randomisation of results.
List state and message.  There is no provision seperate warn or danger messages.  It is just "message".
Removed IP address prefix hardcoded.  The IP address should come from the database.  The dashboard should not make up it's own information.  Just display what is in the database.